+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Defined a LoadContext type
+- We now pass and object into the loader context based on the fetch event.
+
 ## [4.0.2] - 2025-03-08
 
 ### Fixed